test(evals): read split facts through the world instead of raw evaluation
